% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{relief}
\alias{relief}
\title{relief data}
\format{
An object of class \code{data.frame} with 1135311 rows and 3 columns.
}
\usage{
relief
}
\description{
A dataset containing a simplified topographical relief of Switzerland
}
\section{}{
NA
}

\keyword{datasets}
