% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thema_metadata.R
\name{thema_geom}
\alias{thema_geom}
\title{Get the available codes for a given category}
\usage{
thema_geom(category = c("inst", "rrep", "anal", "rtyp", "inke", "topo"))
}
\arguments{
\item{category}{A string corresponding to the category code}
}
\value{
A datafram with code as well as French and German descriptions
}
\description{
Get the available codes for a given category
}
