% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thema_data.R
\name{thema_map}
\alias{thema_map}
\title{Get the map data for a given category, geometry and year}
\usage{
thema_map(
  category = c("inst", "rrep", "anal", "rtyp", "inke", "topo"),
  geometry,
  year,
  level = c("gf", "vf")
)
}
\arguments{
\item{category}{A string corresponding to the category code}

\item{geometry}{A string corresponding to the geometry code}

\item{year}{An integer corresponding to the year of the geometry}

\item{level}{One of gf (general) or vf (limited surfaces)}
}
\value{

}
\description{
Get the map data for a given category, geometry and year
}
