% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thema_data.R
\name{thema_topo}
\alias{thema_topo}
\title{Get topo data}
\usage{
thema_topo(geometry = c("flus", "seen", "stkt"))
}
\arguments{
\item{geometry}{A string corresponding to the geometry code}
}
\value{

}
\description{
Get topo data
}
