% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thema_metadata.R
\name{thema_year}
\alias{thema_year}
\title{Get the years available for a given geometry}
\usage{
thema_year(
  category = c("inst", "rrep", "anal", "rtyp", "inke", "topo"),
  geometry
)
}
\arguments{
\item{category}{A string corresponding to the category code}

\item{geometry}{A string corresponding to the geometry code}
}
\value{
A vector of years
}
\description{
Get the years available for a given geometry
}
